Move `Client` and `Server` to `tcp` namespace
`send` and `receive` moved to separate files so that they can be shared
between `Client` and `Server`.
