+++ tcp/send.hpp
@@ -6,16 +6,17 @@
 
 namespace tcp {
 class SendException : public std::exception {
-  const char* what() const noexcept {
-    return std::string("Could not send message: " + message).c_str();
-  }
+  std::string message;
 
  public:
   SendException(const std::string& in_message) : message(in_message){};
 
- private:
-  std::string message;
+  const char* what() const noexcept {
+    return std::string("Could not send message: " + message).c_str();
+  }
 };
+
+ssize_t send(int fd, const std::string message);
 }  // namespace tcp
 
 #endif
